Added an exception handler around the test runner, but still print the error. Also print a user-friendly explaination of what might have gone wrong.
This solves the problem of students having cryptic tracebacks. They are cryptic because the test script fails, usually when calling a non-existing function, or a type-error. Because the traceback includes code from the wincpy check, it looks as if wincpy is bugged/broken, but actually the user did something wrong.

Possible improvements:
- Command-line option to print the whole traceback, this can be usefull if the problem is actually in the test, and we need a bug report from a user.
- Option to print the whole traceback in case of testing the test runners.